Move TypeNameHelper to util folder
