+++ src/main/services/dataCollectors/RevisionsFilesCollector.groovy
@@ -3,7 +3,7 @@ package services.dataCollectors
 import interfaces.DataCollector
 import project.MergeCommit
 import project.Project
-import services.TypeNameHelper
+import util.TypeNameHelper
 import util.FileManager
 
 import static app.MiningFramework.arguments
